removed unnecessary casts
git-svn-id: http://svn.codehaus.org/groovy/trunk/groovy/groovy-core@792 a5544e8c-8a19-0410-ba12-f9af4593a198

+++ src/main/org/codehaus/groovy/syntax/lexer/Lexer.java
@@ -732,7 +732,7 @@ public class Lexer {
                         LITERAL_LOOP : while (true) {
                             c = la();
 
-                            int x = (int) c;
+                            int x = c;
 
                             LITERAL_SWITCH : switch (c) {
                                 case ('\\') :
